mangoapp: split proto details into separate header
