+++ src/app/main.cpp
@@ -3,6 +3,9 @@
 // If you are new to Dear ImGui, read documentation from the docs/ folder + read the top of imgui.cpp.
 // Read online: https://github.com/ocornut/imgui/tree/master/docs
 
+#include <sys/ipc.h>
+#include <sys/msg.h>
+
 #include "imgui.h"
 #include "imgui_impl_glfw.h"
 #include "imgui_impl_opengl3.h"
@@ -12,6 +15,7 @@
 #include "../overlay.h"
 #include "notify.h"
 #include "mangoapp.h"
+#include "mangoapp_proto.h"
 #include <GL/glew.h>
 #include <GLFW/glfw3.h>
